source/requirements.txt
flask
flask_cors
strawberry-graphql
